This (somewhat) addresses this issue: https://github.com/alinke/PIXEL/issues/8
The server side code now stores a bit of information on what the last upload was, animation or still image.

Now when the page loads, it asks the server side what the last upload was (still, animation, or upload processed).  Then the page restores the GUI for the appropriate upload type or back to scrolling text (if an upload was not the last action).